wip: move the login request to a request package, and make refresh request
