+++ malime.kitsu/src/main/java/com/chesire/malime/kitsu/models/request/LoginRequest.kt
@@ -1,4 +1,4 @@
-package com.chesire.malime.kitsu.models
+package com.chesire.malime.kitsu.models.request
 
 import com.google.gson.annotations.SerializedName
